link files to actual storage

+++ controllers/users-controller.js
@@ -3,12 +3,10 @@ const multer = require("multer");
 const path = require("path");
 const queries = require("../database/queries");
 const { body } = require("express-validator");
+const supabase = require("../database/supabaseClient");
 
 // Storage handling
-const storage = multer.diskStorage({
-  destination: function (req, file, cb) {
-    cb(null, path.resolve("./tmp/uploads"));
-  },
+const storage = multer.memoryStorage({
   filename: function (req, file, cb) {
     const uniqueSuffix = Date.now() + "-" + Math.round(Math.random() * 1e9);
     cb(null, file.fieldname + "-" + uniqueSuffix + file.originalname);
@@ -101,20 +99,28 @@ exports.uploadPost = [
     let path = req.params.path;
     if (path === "root") path = "";
     const file = req.file;
-    await queries.registerFile(
-      file.originalname,
-      file.filename,
-      file.path,
-      file.size,
-      path,
-      user.id
-    );
-    /*
-    name: originalname,
-    storedName: filename,
-    path: path,
-    size: size
-    */
+    file.path = `${user.id}/${Date.now()}_${file.originalname}`;
+    try {
+      const { data, error } = await supabase.storage
+        .from("fileUploader")
+        .upload(file.path, file.buffer, { contentType: file.mimetype });
+      await queries.registerFile(
+        file.originalname,
+        "N/A",
+        data.path,
+        file.size,
+        path,
+        user.id
+      );
+      /*
+      name: originalname,
+      storedName: filename,
+      path: path,
+      size: size
+      */
+    } catch (error) {
+      console.log(error);
+    }
     if (path === "") path = "root";
     res.redirect(`/users/dashboard/${path}`);
   },
@@ -126,6 +132,11 @@ exports.deleteFilePost = async (req, res) => {
     let name = req.params.name;
     let current = req.body.currentPath;
     let path = current + "/" + name;
+    const file = await queries.getFileByPath(path, user.id);
+    if (!file.path.startsWith(`${user.id}/`)) {
+      return res.status(403).send("Forbidden");
+    }
+    await supabase.storage.from("fileUploader").remove([file.path]);
     await queries.deleteFile(path, user.id);
     current = current.replace("/", "%2F").replace(" ", "%20");
     res.redirect(`/users/dashboard/${current}`);
@@ -164,7 +175,13 @@ exports.download = async (req, res) => {
   if (user) {
     const id = +req.params.id;
     const file = await queries.getFileById(id, user.id);
-    res.download(file.path, file.name);
+    if (!file.path.startsWith(`${user.id}`)) {
+      return res.status(403).send("Forbidden");
+    }
+    const { data, error } = await supabase.storage
+      .from("fileUploader")
+      .createSignedUrl(file.path, 60, { download: file.name });
+    res.redirect(data.signedUrl);
   } else {
     res.redirect("/auth");
   }
